requirements.txt
jaclang
jac-cloud
streamlit
mtllm
requests
streamlit
websockets
langchain-community
langchain-text-splitters
langchain-openai
langchain-chroma
streamlit-autorefresh
